+++ benchmarks/src/main/java/com/ionutbalosin/jvm/performance/benchmarks/macro/crypto/util/CryptoUtils.java
@@ -1,0 +1,10 @@
+package com.ionutbalosin.jvm.performance.benchmarks.macro.crypto.util;
+/*
+ * (c) 2023 Ionut Balosin
+ * Website: www.ionutbalosin.com
+ * Twitter: @ionutbalosin
+ *
+ * For the full copyright and license information, please view the LICENSE file that was distributed with this source code.
+ */
+ public class CryptoUtils {
+}
